move ui and clicommon modules up a dir because they are shared now
