+++ lib/clicommon.js
