+++ lib/hotpatch-rebalancer-agent/do_deploy.js
@@ -22,8 +22,7 @@ var vasync = require('vasync');
 var VError = require('verror');
 
 var common = require('../common');
-// XXX
-var clicommon = require('../mantav2-migrate/clicommon');
+var clicommon = require('../clicommon');
 
 function do_deploy(subcmd, opts, args, cb) {
     let hotpatcher = this.hotpatcher;
