+++ lib/hotpatch-rebalancer-agent/index.js
@@ -14,10 +14,9 @@ var cmdln = require('cmdln');
 var restifyClients = require('restify-clients');
 var uuidv4 = require('uuid/v4');
 
-// XXX imports from mantav2-migrate should move up a dir
-var clicommon = require('../mantav2-migrate/clicommon');
+var clicommon = require('../clicommon');
 var Hotpatcher = require('./hotpatcher').Hotpatcher;
-var UI = require('../mantav2-migrate/ui').UI;
+var UI = require('../ui').UI;
 
 // ---- globals
