+++ lib/mantav2-migrate/do_snaplink_cleanup.js
@@ -22,7 +22,7 @@ var cmdln = require('cmdln');
 var vasync = require('vasync');
 var VError = require('verror').VError;
 
-var clicommon = require('./clicommon');
+var clicommon = require('../clicommon');
 
 var promptYesNo = clicommon.promptYesNo;
