+++ lib/mantav2-migrate/index.js
@@ -14,9 +14,9 @@ var cmdln = require('cmdln');
 var restifyClients = require('restify-clients');
 var uuidv4 = require('uuid/v4');
 
-var clicommon = require('./clicommon');
+var clicommon = require('../clicommon');
 var Migrator = require('./migrator').Migrator;
-var UI = require('./ui').UI;
+var UI = require('../ui').UI;
 
 // ---- globals
