+++ lib/ui.js
@@ -10,10 +10,9 @@
  * Note: This is adapted from <sdcadm.git:lib/cli/ui.js>.
  *
  * A class to capture all/most output on the CLI. Typically the CLI object
- * will have an instance of this at `this.ui`. `mbucket` code should use that
- * to write output for the user.
+ * will have an instance of this at `this.ui`.
  *
- * Usage in mantav2-migrate code:
+ * Usage in code for a CLI tool using this:
  *
  * - get the `<cli>.ui` object
  * - `ui.info(...)` for printf-style message output to stdout
